+++ tests/specs/listenTo.spec.js
@@ -1,0 +1,267 @@
+var Rebase = require('../../dist/bundle');
+var React = require('react');
+var ReactDOM = require('react-dom');
+var firebase = require('firebase');
+
+var invalidEndpoints = require('../fixtures/invalidEndpoints');
+var dummyObjData = require('../fixtures/dummyObjData');
+var invalidOptions = require('../fixtures/invalidOptions');
+var firebaseConfig = require('../fixtures/config');
+
+describe('listenTo()', function(){
+  var base;
+  var testEndpoint = 'test/listenTo';
+  var testApp;
+  var ref;
+
+  beforeAll(() => {
+    testApp = firebase.initializeApp(firebaseConfig, 'DB_CHECK');
+    ref = testApp.database().ref();
+    var mountNode = document.createElement('div');
+    mountNode.setAttribute("id", "mount");
+    document.body.appendChild(mountNode);
+  });
+
+  afterAll(done => {
+    testApp.delete().then(done);
+  });
+
+  beforeEach(() => {
+    base = Rebase.createClass(firebaseConfig);
+  });
+
+  afterEach(done => {
+    ReactDOM.unmountComponentAtNode(document.body);
+    firebase.Promise.all([
+      base.delete(),
+      ref.child(testEndpoint).set(null)
+    ]).then(done);
+  });
+
+  it('listenTo() returns a valid ref', function(){
+    var ref = base.listenTo(testEndpoint, {
+      context: this,
+      then(data){}
+    });
+    expect(ref.id).toEqual(jasmine.any(Number));
+  })
+
+  it('listenTo() throws an error given a invalid endpoint', function(){
+    invalidEndpoints.forEach((endpoint) => {
+      try {
+        base.listenTo(endpoint, {
+          context: this,
+          then(data){
+          }
+        })
+      } catch(err) {
+        expect(err.code).toEqual('INVALID_ENDPOINT');
+      }
+    });
+  });
+
+  it('listenTo() throws an error given an invalid options object', function(){
+    invalidOptions.forEach((option) => {
+      try {
+        base.listenTo(testEndpoint, option);
+      } catch(err) {
+        expect(err.code).toEqual('INVALID_OPTIONS');
+      }
+    });
+  });
+
+  describe('Async tests', function(){
+    it('listenTo()\'s .then method gets invoked when the Firebase endpoint changes', function(done){
+      var didUpdate = false;
+      base.listenTo(testEndpoint, {
+        context: {},
+        then(data){
+          if(didUpdate === true){
+            expect(data).toEqual(dummyObjData);
+            done()
+          }
+        }
+      });
+      ref.child(testEndpoint).set(dummyObjData).then(() => {
+        didUpdate = true;
+      });
+    });
+
+    it('listenTo\'s .then method gets invoked when the Firebase endpoint changes and correctly updates the component\'s state', function(done){
+      var didUpdate = false;
+      class TestComponent extends React.Component{
+        constructor(props){
+          super(props);
+          this.state = {
+            data: {}
+          }
+        }
+        componentWillMount(){
+          this.ref = base.listenTo(testEndpoint, {
+            context: this,
+            then(data){
+              this.setState({data})
+            }
+          });
+        }
+        componentDidMount(){
+          ref.child(testEndpoint).set(dummyObjData).then(() => {
+            didUpdate = true;
+          });
+        }
+        componentDidUpdate(){
+          if(didUpdate){
+            expect(this.state.data).toEqual(dummyObjData);
+            done();
+          }
+        }
+        render(){
+          return (
+            <div>
+              Name: {this.state.name} <br />
+              Age: {this.state.age}
+            </div>
+          )
+        }
+      }
+      ReactDOM.render(<TestComponent />, document.getElementById('mount'));
+    });
+
+
+    it('listenTo should return the data as an array if the asArray property of options is set to true', function(done){
+      var didUpdate = false;
+      class TestComponent extends React.Component{
+        constructor(props){
+          super(props);
+          this.state = {
+            data: {}
+          }
+        }
+        componentWillMount(){
+          this.ref = base.listenTo(testEndpoint, {
+            context: this,
+            then(data){
+              this.setState({data});
+            },
+            asArray: true
+          });
+        }
+        componentDidMount(){
+          ref.child(testEndpoint).set(dummyObjData).then(() => {
+            didUpdate = true;
+          });
+        }
+        componentDidUpdate(){
+          if(didUpdate){
+            expect(this.state.data.indexOf(25)).not.toBe(-1);
+            expect(this.state.data.indexOf('Tyler McGinnis')).not.toBe(-1);
+            done();
+          }
+        }
+        render(){
+          return (
+            <div>
+              Name: {this.state.name} <br />
+              Age: {this.state.age}
+            </div>
+          )
+        }
+      }
+      ReactDOM.render(<TestComponent />, document.getElementById('mount'));
+    });
+
+    it('listenTo should allow multiple components to listen to changes on the same endpoint', function(done){
+      //set up mount points
+      var div1 = document.createElement('div');
+      div1.setAttribute("id", "div1");
+
+      var div2 = document.createElement('div');
+      div2.setAttribute("id", "div2");
+
+      document.body.appendChild(div1);
+      document.body.appendChild(div2);
+
+      function cleanUp(done){
+        ReactDOM.unmountComponentAtNode(document.getElementById('div1'));
+        ReactDOM.unmountComponentAtNode(document.getElementById('div2'));
+        done();
+      }
+
+      var component1Updated = false;
+      var component2Updated = false;
+      
+      class TestComponent1 extends React.Component{
+        constructor(props){
+          super(props);
+          this.state = {
+            data: {}
+          }
+        }
+        componentDidMount(){
+          this.ref = base.listenTo(testEndpoint, {
+            context: this,
+            then(data){
+              this.setState({data});
+            },
+            asArray: true
+          });
+        }
+        componentDidUpdate(){
+          expect(this.state.data.indexOf(25)).not.toBe(-1);
+          expect(this.state.data.indexOf('Tyler McGinnis')).not.toBe(-1);
+          component1Updated = true;
+          if(component1Updated && component2Updated){
+            cleanUp(done);
+          }
+        }
+        render(){
+          return (
+            <div>
+              Name: {this.state.name} <br />
+              Age: {this.state.age}
+            </div>
+          )
+        }
+      }
+      class TestComponent2 extends React.Component{
+        constructor(props){
+          super(props);
+          this.state = {
+            data: {}
+          }
+        }
+        componentWillMount(){
+          this.ref = base.listenTo(testEndpoint, {
+            context: this,
+            then(data){
+              this.setState({data});
+            },
+            asArray: true
+          });
+        }
+        componentDidMount(){
+          ref.child(testEndpoint).set(dummyObjData)
+        }
+        componentDidUpdate(){
+          expect(this.state.data.indexOf(25)).not.toBe(-1);
+          expect(this.state.data.indexOf('Tyler McGinnis')).not.toBe(-1);
+          component2Updated = true;
+          if(component1Updated && component2Updated){
+            cleanUp(done);
+          }
+        }
+        render(){
+          return (
+            <div>
+              Name: {this.state.name} <br />
+              Age: {this.state.age}
+            </div>
+          )
+        }
+      }
+      ReactDOM.render(<TestComponent1 />, document.getElementById('div1'));
+      ReactDOM.render(<TestComponent2 />, document.getElementById('div2'));
+    });
+  });
+
+});
